# perf(osmomath): make Exp2 use cheaper comparisons firstly