This change moves the cheaper big.Int.IsNegative() before big.Int.Abs().GT(maxSupportedExponent) to quickly throw away negative exponent values much faster than firstly converting to an absolute value then comparing against a large value. This change also includes a newly added benchmark that clearly exhibits the improvement.

```shell
$ benchstat before.txt after.txt
name             old time/op    new time/op    delta
Exp2Negatives-8    12.7µs ± 2%     9.7µs ± 3%  -23.04%  (p=0.000 n=10+10)

name             old alloc/op   new alloc/op   delta
Exp2Negatives-8    6.16kB ± 0%    4.35kB ± 0%  -29.35%  (p=0.000 n=9+10)

name             old allocs/op  new allocs/op  delta
Exp2Negatives-8       119 ± 0%        88 ± 0%  -26.05%  (p=0.000 n=10+10)
```

Fixes #7740

<!-- This is an auto-generated comment: release notes by coderabbit.ai -->

## Summary by CodeRabbit

- **Refactor**
	- Improved the efficiency of the `Exp2` function by adjusting the order of exponent magnitude checks.
- **Tests**
	- Added benchmark tests for negative exponents in the `Exp2` function.

<!-- end of auto-generated comment: release notes by coderabbit.ai -->